edit menu constant
