+++ src/components/Dashboard/Dashboard.constants.js
@@ -1,0 +1,44 @@
+const menu = {
+  categories: [
+    {
+      name: "tatli",
+      items: [
+        {
+          name: "su",
+          price: "9",
+        },
+        {
+          name: "lo",
+          price: "1",
+        },
+      ],
+    },
+    {
+      name: "icecek",
+      items: [
+        {
+          name: "su",
+          price: "9",
+        },
+        {
+          name: "sldsassjdflo",
+          price: "1",
+        },
+      ],
+    },
+    {
+      name: "icecek",
+      items: [
+        {
+          name: "su",
+          price: "9",
+        },
+        {
+          name: "sldsassjdflo",
+          price: "1",
+        },
+      ],
+    },
+  ],
+};
+export default menu;
