+++ src/components/Dashboard/Dashboard.js
@@ -3,55 +3,14 @@ import React from 'react';
 import { useNavigate } from "react-router-dom";
 import { CategoryTable } from '../../elements/menuElements';
 import Header from '../Header/Header';
-
+import menu from "./Dashboard.constants";
 const Dashboard = () => {
-  const navigate = useNavigate()
-  const user = JSON.parse(localStorage.getItem('loggedInUser'))
+  const navigate = useNavigate();
+  const user = JSON.parse(localStorage.getItem("loggedInUser"));
 
   if (!user) {
-    navigate('/login')
+    navigate("/login");
   }
-  const menu = [
-    {
-      categoryName: "tatli",
-      items: [
-        {
-          itemName: "su",
-          itemPrice: "9",
-        },
-        {
-          itemName: "lo",
-          itemPrice: "1",
-        },
-      ],
-    },
-    {
-      categoryName: "icecek",
-      items: [
-        {
-          itemName: "su",
-          itemPrice: "9",
-        },
-        {
-          itemName: "sldsassjdflo",
-          itemPrice: "1",
-        },
-      ],
-    },
-    {
-      categoryName: "icecek",
-      items: [
-        {
-          itemName: "su",
-          itemPrice: "9",
-        },
-        {
-          itemName: "sldsassjdflo",
-          itemPrice: "1",
-        },
-      ],
-    },
-  ];
 
   return (
     <Flex direction="column" gap="12px" paddingBottom="20px">
@@ -62,13 +21,13 @@ const Dashboard = () => {
           Menu
         </Heading>
         <Flex direction="column" flexWrap="wrap" gap="16px" alignItems="center">
-          {menu.map((category, index) => {
-            return <CategoryTable key={index} category={category} />;
-          })}
+          {menu.categories.map((category, index) => (
+            <CategoryTable key={index} category={category} />
+          ))}
         </Flex>
       </Box>
     </Flex>
   );
-}
+};
 
 export default Dashboard;
