+++ src/components/SignUpForm/SignUpForm.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import { FormContainer, PasswordInput, SubmitButton, TextInput } from '../../elements/formElements';
+import { SubmitButton } from "../../elements/buttons";
+import { FormContainer, PasswordInput, TextInput } from "../../elements/formElements";
 import useField from '../../hooks/useField';
 import signupService from '../../services/signup';
