+++ src/elements/buttons.js
@@ -48,3 +48,20 @@ export const CancelButton = ({ onClick }) => {
     </Button>
   );
 };
+
+export const SubmitButton = ({ buttonText }) => {
+  return (
+    <Button
+      type="submit"
+      borderRadius="20px"
+      bg="#6D8B74"
+      color="white"
+      w="full"
+      _hover={{
+        bg: "#5F7161",
+      }}
+    >
+      {buttonText}
+    </Button>
+  );
+};
